+++ jody_hash.h
@@ -22,7 +22,7 @@ typedef uint64_t hash_t;
 #define JODY_HASH_VERSION 3
 
 extern hash_t jody_block_hash(const hash_t * restrict data,
-		const hash_t start_hash, const unsigned int count);
+		const hash_t start_hash, const size_t count);
 
 #ifdef __cplusplus
 }
